chore: remove or change Angular version number
This commit removes the version number "2" of all of our articles where possible.
In some cases, simply removing the version number isn't enough as it can be confusing
when e.g. talking about upgrading from Angular 1.x to Angular >= 2.x.

Based on context, instead of removing "2", this commit adds `>= 2.x` or changes
sentences to sth. like "...since version 2.x, Angular does...".

This commit also adds a information box at the beginning of the every article that
covers Angular >= 2.x to make sure the reader knows which version we're dealing with,
because it's not obvious when reading the post title anymore.